+++ app/src/main/java/com/fpt/team5/golddigger/LoginActivity.java
@@ -74,7 +74,7 @@ public class LoginActivity extends AppCompatActivity {
                 do {
                     int id = c.getInt(0);
                     String title = c.getString(1);
-                    float amount = c.getFloat(2);
+                    double amount = c.getDouble(2);
                     int uId = c.getInt(3);
                     String date = c.getString(4);
